+++ test/index.js
@@ -110,8 +110,14 @@ describe('Wallet fixtures', function () {
 
       describe('fromJSON', function () {
         it('imports from a JSON object', function () {
-          assert.equal(wallet.account.chains[0].addresses, f.json.external.addresses)
-          assert.equal(wallet.account.chains[1].addresses, f.json.internal.addresses)
+          Object.keys(f.json.external.map).forEach(function (address) {
+            assert(wallet.account.chains[0].addresses.indexOf(address) !== -1, address)
+          })
+
+          Object.keys(f.json.internal.map).forEach(function (address) {
+            assert(wallet.account.chains[1].addresses.indexOf(address) !== -1, address)
+          })
+
           assert.equal(wallet.account.chains[0].map, f.json.external.map)
           assert.equal(wallet.account.chains[1].map, f.json.internal.map)
           assert.equal(wallet.unspents, f.json.unspents)
@@ -120,11 +126,11 @@ describe('Wallet fixtures', function () {
 
       describe('containsAddress', function () {
         it('returns the expected results', function () {
-          f.json.external.addresses.forEach(function (address) {
+          Object.keys(f.json.external.map).forEach(function (address) {
             assert(wallet.containsAddress(address))
           })
 
-          f.json.internal.addresses.forEach(function (address) {
+          Object.keys(f.json.internal.map).forEach(function (address) {
             assert(wallet.containsAddress(address))
           })
 
@@ -241,9 +247,12 @@ describe('Wallet fixtures', function () {
 
       describe('getAllAddresses', function () {
         it('returns all known addresses', function () {
-          var addresses = f.json.external.addresses.concat(f.json.internal.addresses)
+          var fAllAddresses = Object.keys(f.json.external.map).concat(Object.keys(f.json.internal.map))
+          var allAddresses = wallet.getAllAddresses()
 
-          assert.deepEqual(wallet.getAllAddresses(), addresses)
+          fAllAddresses.forEach(function (address) {
+            assert(allAddresses.indexOf(address) !== -1, address)
+          })
         })
       })