+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 # 0.4.0
 * removes `Wallet.getConfirmedBalance`, `confirmations` no longer used in the API, you must now filter unspents yourself prior to use
+* removes `.addresses` from chain JSON,  was deterministic